fix: fixed issues after PR
