+++ core/src/main/java/org/openedx/core/AppDataConstants.kt
@@ -11,5 +11,6 @@ object AppDataConstants {
     const val VIDEO_FORMAT_M3U8 = ".m3u8"
     const val VIDEO_FORMAT_MP4 = ".mp4"
 
-    const val DOWNLOADS_CONFIRMATION_SIZE = 1073741824L
+    // Equal 1GB
+    const val DOWNLOADS_CONFIRMATION_SIZE = 1024 * 1024 * 1024L
 }
