--- core/src/main/java/org/openedx/core/system/notifier/ProfileEvent.kt
@@ -1,3 +1,0 @@
-package org.openedx.core.system.notifier
-
-interface ProfileEvent
