+++ core/src/main/java/org/openedx/core/system/notifier/VideoEvent.kt
@@ -1,3 +1,3 @@
 package org.openedx.core.system.notifier
 
-class AccountDeactivated : ProfileEvent
+interface VideoEvent
