+++ profile/src/main/java/org/openedx/profile/system/notifier/AccountDeactivated.kt
@@ -1,0 +1,3 @@
+package org.openedx.profile.system.notifier
+
+class AccountDeactivated : ProfileEvent
