+++ profile/src/main/java/org/openedx/profile/system/notifier/AccountUpdated.kt
@@ -1,0 +1,3 @@
+package org.openedx.profile.system.notifier
+
+class AccountUpdated : ProfileEvent
