+++ profile/src/main/java/org/openedx/profile/system/notifier/ProfileEvent.kt
@@ -1,0 +1,3 @@
+package org.openedx.profile.system.notifier
+
+interface ProfileEvent
