+++ profile/src/main/java/org/openedx/profile/system/notifier/ProfileNotifier.kt
@@ -1,8 +1,9 @@
-package org.openedx.core.system.notifier
+package org.openedx.profile.system.notifier
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import org.openedx.core.system.notifier.VideoQualityChanged
 
 class ProfileNotifier {
 
@@ -11,7 +12,6 @@ class ProfileNotifier {
     val notifier: Flow<ProfileEvent> = channel.asSharedFlow()
 
     suspend fun send(event: AccountUpdated) = channel.emit(event)
-    suspend fun send(event: VideoQualityChanged) = channel.emit(event)
     suspend fun send(event: AccountDeactivated) = channel.emit(event)
 
 }
