+++ profile/src/test/java/org/openedx/profile/presentation/profile/ProfileViewModelTest.kt
@@ -34,8 +34,8 @@ import org.openedx.core.system.notifier.AppUpgradeNotifier
 import org.openedx.profile.domain.interactor.ProfileInteractor
 import org.openedx.profile.presentation.ProfileAnalytics
 import org.openedx.profile.presentation.ProfileRouter
-import org.openedx.core.system.notifier.AccountUpdated
-import org.openedx.core.system.notifier.ProfileNotifier
+import org.openedx.profile.system.notifier.AccountUpdated
+import org.openedx.profile.system.notifier.ProfileNotifier
 import java.net.UnknownHostException
 
 @OptIn(ExperimentalCoroutinesApi::class)
